Removed all the hey-api code we are not using currently.
Hey API is the way Umbraco have suggested people can consume the Swagger API's and produced strongly typed TypeScript models. However they have depreciated the way Umbraco suggest it should be used and I can't find any examples ( or get it working! ) of the new Hey API way of doing it..

For now, this is using a hybrid approach, it's using the Types generated by Hey API, but doing "manual" fetches calling the API's,

I have heard good things about Orval as an alternative, so it might be a good idea to look into that.